# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_handlers.ExchangeRequest:
    properties:
      amount:
        type: number
      from_currency:
        type: string
      to_currency:
        type: string
    required:
    - amount
    - from_currency
    - to_currency
    type: object
  internal_handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_handlers.LoginResponse:
    properties:
      token:
        type: string
    type: object
  internal_handlers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  internal_handlers.WalletOperation:
    properties:
      amount:
        type: number
      currency:
        enum:
        - USD
        - RUB
        - EUR
        type: string
    required:
    - amount
    - currency
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for currency wallet operations
  title: Currency Wallet API
  version: "1.0"
paths:
  /balance:
    get:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user total balance for all currencies
      tags:
      - wallet
  /balance/{currency}:
    get:
      parameters:
      - description: Currency code (USD, RUB, EUR)
        in: path
        name: currency
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user balance for currency
      tags:
      - wallet
  /exchange:
    post:
      consumes:
      - application/json
      parameters:
      - description: Exchange request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ExchangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Exchange currencies
      tags:
      - exchange
  /exchange/rates:
    get:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get current exchange rates
      tags:
      - exchange
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login and get JWT token
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /wallet/deposit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Deposit request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.WalletOperation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Deposit funds to wallet
      tags:
      - wallet
  /wallet/withdraw:
    post:
      consumes:
      - application/json
      parameters:
      - description: Withdraw request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.WalletOperation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Withdraw funds from wallet
      tags:
      - wallet
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"